Move DSL component classes into separate directory, also;
- Remove deprecated methods
- Add more documentation
- Whitespace changes
