+++ lib/vedeu/dsl/all.rb
@@ -1,14 +1,14 @@
+require 'vedeu/dsl/components/all'
 require 'vedeu/dsl/shared/all'
 
-require 'vedeu/dsl/border'
 require 'vedeu/dsl/composition'
-require 'vedeu/dsl/geometry'
 require 'vedeu/dsl/interface'
-require 'vedeu/dsl/keymap'
 require 'vedeu/dsl/line'
-require 'vedeu/dsl/menu'
 require 'vedeu/dsl/stream'
 require 'vedeu/dsl/view'
 
 module Vedeu
+
+
+
 end # Vedeu
