+++ lib/vedeu/dsl/components/all.rb
@@ -1,0 +1,10 @@
+require 'vedeu/dsl/components/border'
+require 'vedeu/dsl/components/geometry'
+require 'vedeu/dsl/components/keymap'
+require 'vedeu/dsl/components/menu'
+
+module Vedeu
+
+
+
+end # Vedeu
