+++ lib/vedeu/dsl/components/border.rb
