+++ lib/vedeu/dsl/components/geometry.rb
