+++ lib/vedeu/dsl/components/menu.rb
