+++ lib/vedeu/dsl/shared/all.rb
@@ -3,4 +3,7 @@ require 'vedeu/dsl/shared/colour'
 require 'vedeu/dsl/shared/style'
 
 module Vedeu
+
+
+
 end # Vedeu
