+++ test/lib/vedeu/dsl/components/border_test.rb
