+++ test/lib/vedeu/dsl/components/geometry_test.rb
