+++ test/lib/vedeu/dsl/components/keymap_test.rb
