+++ test/lib/vedeu/dsl/components/menu_test.rb
